requirements.txt
alembic==1.0.10
aniso8601==6.0.0
attrs==19.1.0
awsebcli==3.12.2
bcrypt==3.1.6
botocore==1.8.44
cement==2.8.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudant==2.4.0
colorama==0.3.7
coverage==4.4.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum-compat==0.0.2
eventlet==0.21.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
gem==0.1.12
greenlet>=0.4.12
gunicorn==19.7.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==3.0.1
Mako==1.0.9
MarkupSafe==1.1.1
pathspec==0.5.5
psycopg2==2.7.3.2
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-engineio==2.0.2
python-socketio==1.8.4
pytz==2019.1
PyYAML>=3.12
requests==2.21.0
selenium==3.8.1
semantic-version==2.5.0
six==1.12.0
SQLAlchemy==1.3.3
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.24.2
virtualenv==16.5.0
websocket-client==0.46.0
Werkzeug==0.15.2
WTForms==2.2.1
WTForms-JSON==0.3.3
